+++ src/@types/casperlabsSigner.d.ts
@@ -1,5 +1,10 @@
 interface CasperLabsHelper {
   /**
+   * Returns Signer version
+   */
+  getVersion: () => Promise<string>;
+
+  /**
    * Returns connection status from Signer
    */
   isConnected: () => Promise<boolean>;
@@ -10,13 +15,14 @@ interface CasperLabsHelper {
   requestConnection: () => void;
 
   /**
-   * send base16 encoded message to plugin to sign
+   * Send Deploy in JSON format message to Signer plugin to sign.
    *
-   * @param messageBase16 the base16 encoded message that plugin received to sign
-   * @param publicKeyBase64 the base64 encoded public key used to sign the deploy, if set, we will check whether it is the same as the active key for signing the message, otherwise, we won't check.
+   * @param deploy deploy in JSON format
+   * @param sourcePublicKeyHex public key in hex format with algorithm prefix. Used to sign the deploy
+   * @param targetPublicKeyHex public key in hex format with algorithm prefix. Used to display hex-formatted address on the UI
    */
-  sign: (messageBase16: string, publicKeyBase64?: string) => Promise<string>;
-
+  sign: (deploy: any, sourcePublicKeyHex: string, targetPublicKeyHex: string) => Promise<JSON>;
+  
   /*
    * Returns base64 encoded public key of user current selected account.
    */