+++ src/lib/Signer.ts
@@ -6,6 +6,17 @@
  */
 
 /**
+ * Returns Signer version
+ */
+export const getVersion: () => Promise<string> = async () => {
+  try {
+    return await window.casperlabsHelper!.getVersion();
+  } catch {
+    return "<1.0.0";
+  }
+};
+
+/**
  * Check whether CasperLabs Signer extension is connected
  */
 export const isConnected: () => Promise<boolean> = async () => {
@@ -38,19 +49,24 @@ export const getActivePublicKey: () => Promise<string> = () => {
 };
 
 /**
- * send base16 encoded message to plugin to sign
+ * Send Deploy in JSON format message to Signer plugin to sign.
  *
- * @param messageBase16 the base16 encoded message that plugin received to sign
- * @param publicKeyBase64 the base64 encoded public key used to sign the deploy, if set, we will check whether it is the same as the active key for signing the message, otherwise, we won't check.
+ * @param deploy deploy in JSON format
+ * @param sourcePublicKeyHex public key in hex format with algorithm prefix. Used to sign the deploy
+ * @param targetPublicKeyHex public key in hex format with algorithm prefix. Used to display hex-formatted address on the UI
  *
  * @throws Error if haven't connected to CasperLabs Signer browser extension.
- * @throws Error if publicKeyBase64 is not the same as the key that Signer used to sign the message
+ * @throws Error if sourcePublicKeyHex is not the same as the key that Signer used to sign the message
+ * @throws Error if targetPublicKeyHex is not the same as the key that is used as target in deploy.
  */
 export const sign: (
-  messageBase16: string,
-  publicKeyBase64?: string
-) => Promise<string> = (messageBase16: string, publicKeyBase64?: string) => {
-  return window.casperlabsHelper!.sign(messageBase16, publicKeyBase64);
+  deploy: any, 
+  sourcePublicKey: string, 
+  targetPublicKey: string
+) => Promise<any> = (
+  deploy: any, sourcePublicKey: string, targetPublicKey: string
+) => {
+  return window.casperlabsHelper!.sign(deploy, sourcePublicKey, targetPublicKey);
 };
 
 /*